Move random to internal
